+++ zscaler/zia/dns_gatways.py
@@ -33,7 +33,7 @@ class DNSGatewayAPI(APIClient):
         super().__init__()
         self._request_executor: RequestExecutor = request_executor
 
-    def list_dns_gateways(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_dns_gateways(self, query_params: Optional[dict] = None) -> APIResult[List[DNSGateways]]:
         """
         Returns a list of dns gateways.