+++ zscaler/zpa/customer_dr_tool.py
@@ -33,7 +33,7 @@ class CustomerDRToolVersionAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_latest_dr_tool_versions(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_latest_dr_tool_versions(self, query_params: Optional[dict] = None) -> APIResult[List[CustomerDRToolVersion]]:
         """
         Fetch latest the Customer Support DR Tool Versions sorted by latest filter.