+++ zscaler/zpa/managed_browser_profile.py
@@ -33,7 +33,7 @@ class ManagedBrowserProfileAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_managed_browser_profiles(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_managed_browser_profiles(self, query_params: Optional[dict] = None) -> APIResult[List[ManagedBrowserProfile]]:
         """
         Gets all the managed browser profiles for a customer