+++ zscaler/zpa/microtenants.py
@@ -34,7 +34,7 @@ class MicrotenantsAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_microtenants(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_microtenants(self, query_params: Optional[dict] = None) -> APIResult[List[Microtenant]]:
         """
         Enumerates microtenants in your organization with pagination.
         A subset of microtenants can be returned that match a supported