+++ zscaler/zpa/pra_credential.py
@@ -33,7 +33,7 @@ class PRACredentialAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_credentials(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_credentials(self, query_params: Optional[dict] = None) -> APIResult[List[PrivilegedRemoteAccessCredential]]:
         """
         Returns a list of all privileged remote access credentials.