+++ zscaler/zpa/pra_portal.py
@@ -33,7 +33,7 @@ class PRAPortalAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_portals(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_portals(self, query_params: Optional[dict] = None) -> APIResult[List[PrivilegedRemoteAccessPortal]]:
         """
         Returns a list of all configured PRA portals with pagination support.