+++ zscaler/zpa/user_portal_controller.py
@@ -33,7 +33,7 @@ class UserPortalControllerAPI(APIClient):
         customer_id = config["client"].get("customerId")
         self._zpa_base_endpoint = f"/zpa/mgmtconfig/v1/admin/customers/{customer_id}"
 
-    def list_user_portals(self, query_params: Optional[dict] = None) -> APIResult[dict]:
+    def list_user_portals(self, query_params: Optional[dict] = None) -> APIResult[List[UserPortalController]]:
         """
         Enumerates user portals in an organization with pagination.